imported 20080201 version of NetBSD-current.

+++ src/usr.sbin/sup/source/scan.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: scan.c,v 1.23 2006/12/20 16:33:34 christos Exp $	*/
+/*	$NetBSD: scan.c,v 1.26 2007/12/20 20:15:59 christos Exp $	*/
 
 /*
  * Copyright (c) 1992 Carnegie Mellon University
@@ -83,8 +83,6 @@
  **********************************************************************
  */
 
-#include "libc.h"
-#include "c.h"
 #ifdef HAS_VIS
 #include <vis.h>
 #endif
@@ -98,9 +96,12 @@
 #include <sys/dir.h>
 #endif
 #include <sys/file.h>
+#include <limits.h>
 #include <unistd.h>
 #include "supcdefs.h"
 #include "supextern.h"
+#include "libc.h"
+#include "c.h"
 
 /*************************
  ***    M A C R O S    ***
@@ -706,9 +707,8 @@ listdir(char *name, int always)
 	struct direct *dentry;
 #endif
 	DIR *dirp;
-	char ename[STRINGLENGTH], newname[STRINGLENGTH], filename[STRINGLENGTH];
+	char newname[STRINGLENGTH], filename[STRINGLENGTH];
 	char *p, *newp;
-	int i;
 
 	dirp = opendir(".");
 	if (dirp == 0)
@@ -736,14 +736,15 @@ listdir(char *name, int always)
 			continue;
 		if (strcmp(dentry->d_name, "..") == 0)
 			continue;
-		for (i = 0; i <= MAXNAMLEN && dentry->d_name[i]; i++)
-			ename[i] = dentry->d_name[i];
-		ename[i] = 0;
-		if (*newname)
-			(void) sprintf(filename, "%s/%s", newname, ename);
-		else
-			(void) strcpy(filename, ename);
-		listentry(ename, filename, newname, always);
+		if (*newname) {
+			(void)snprintf(filename, sizeof(filename), "%s/%s",
+			    newname, dentry->d_name);
+		} else {
+			(void)strncpy(filename, dentry->d_name,
+			    sizeof(filename) - 1);
+			filename[sizeof(filename) - 1] = '\0';
+		}
+		listentry(dentry->d_name, filename, newname, always);
 	}
 	closedir(dirp);
 }
@@ -965,18 +966,25 @@ makescanfile(char *scanfile)
 	(void) sprintf(tname, "%s.temp", fname);
 	scanF = fopen(tname, "w");
 	if (scanF == NULL)
-		goaway("Can't write scan file temp %s for %s", tname, collname);
-	fprintf(scanF, "V%d\n", SCANVERSION);
-	(void) Tprocess(listT, recordone, scanF);
-	(void) fclose(scanF);
-	if (rename(tname, fname) < 0)
+		goto out;
+	if (fprintf(scanF, "V%d\n", SCANVERSION) < 0)
+		goto out;
+	if (Tprocess(listT, recordone, scanF) != SCMOK)
+		goto out;
+	if (fclose(scanF) != 0)
+		goto out;
+	if (rename(tname, fname) < 0) {
+		(void)unlink(tname);
 		goaway("Can't change %s to %s", tname, fname);
-	(void) unlink(tname);
+	}
 	tbuf[0].tv_sec = time((time_t *) NULL);
 	tbuf[0].tv_usec = 0;
 	tbuf[1].tv_sec = scantime;
 	tbuf[1].tv_usec = 0;
 	(void) utimes(fname, tbuf);
+	return;
+out:
+	goaway("Can't write scan file temp %s for %s", tname, collname);
 }
 
 static int
@@ -991,13 +999,15 @@ recordone(TREE * t, void *v)
 #endif
 
 	if (t->Tflags & FBACKUP)
-		fprintf(scanF, "B");
+		if (fprintf(scanF, "B") < 0)
+			return SCMERR;
 	if (t->Tflags & FNOACCT)
-		fprintf(scanF, "N");
-	fprintf(scanF, "%o %d %d %s\n",
-	    t->Tmode, t->Tctime, t->Tmtime, fname);
-	(void) Tprocess(t->Texec, recordexec, scanF);
-	return (SCMOK);
+		if (fprintf(scanF, "N") < 0)
+			return SCMERR;
+	if (fprintf(scanF, "%o %d %d %s\n",
+	    t->Tmode, t->Tctime, t->Tmtime, fname) < 0)
+		return SCMERR;
+	return Tprocess(t->Texec, recordexec, scanF);
 }
 
 static int
@@ -1010,7 +1020,8 @@ recordexec(TREE * t, void *v)
 #else
 	char *fname = t->Tname;
 #endif
-	fprintf(scanF, "X%s\n", fname);
+	if (fprintf(scanF, "X%s\n", fname) < 0)
+		return SCMERR;
 	return (SCMOK);
 }
